.old/web/requirements.txt
certifi>=2018.8.24
chardet>=3.0.4
Cython>=0.28.5
dataclasses>=0.6
Django>=2.0.7
django-channels>=0.7.0
django-rest-framework>=0.1.0
django-webpack-loader>=0.6.0
djangorestframework>=3.8.2
docopt>=0.6.2
ExifRead>=2.1.2
idna>=2.7
numpy>=1.15.0
oauthlib>=2.1.0
opencv-python>=3.4.2.17
packbits>=0.6
Pillow>=5.2.0
psd-tools>=1.4
psd-tools2>=1.7.16
pymongo>=3.7.1
pytoshop>=1.2.0
pytz>=2018.5
redis>=2.10.6
requests>=2.19.1
requests-oauthlib>=1.0.0
scikit-learn>=0.19.2
scipy>=1.1.0
six>=1.11.0
sqlparse>=0.2.4
typing>=3.6.6
urllib3>=1.23
